requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.11
Tempita==0.5.2
WTForms==2.1
Werkzeug==0.11.3
argparse==1.2.1
decorator==4.0.6
httplib2==0.9.2
itsdangerous==0.24
pbr==1.8.1
python-instagram==1.3.2
requests==2.9.1
simplejson==3.8.1
six==1.10.0
sqlalchemy-migrate==0.10.0
sqlparse==0.1.18
wsgiref==0.1.2
